feat: follow up events section with config
