+++ app/config_utils.py
@@ -1,0 +1,14 @@
+import json
+import os
+
+def load_config(file_path):
+    with open(file_path) as f:
+        config = json.load(f)
+    return config
+
+def get_custom_config_dropdowns(config_path):
+    path = os.getcwd()
+    dropdown_config_path = path + config_path
+    dropdown_config = load_config(dropdown_config_path)
+    return dropdown_config
+
